refactor: addressed PR comments bb-258
