+++ apps/frontend/src/libs/components/protected-route/libs/constants/constants.ts
@@ -1,1 +1,1 @@
-export { routesWithoutAuthWrapper } from "./routes-with-auth-wrapper.constant.js";
+export { ROUTES_WITHOUT_AUTH_WRAPPER } from "./routes-without-auth-wrapper.constant.js";
