--- apps/frontend/src/libs/components/protected-route/libs/constants/routes-with-auth-wrapper.constant.ts
@@ -1,6 +1,0 @@
-import { AppRoute } from "~/libs/enums/enums.js";
-import { type ValueOf } from "~/libs/types/types.js";
-
-const routesWithoutAuthWrapper = [AppRoute.QUIZ] as ValueOf<typeof AppRoute>[];
-
-export { routesWithoutAuthWrapper };
