+++ apps/frontend/src/libs/components/protected-route/libs/constants/routes-without-auth-wrapper.constant.ts
@@ -1,0 +1,8 @@
+import { AppRoute } from "~/libs/enums/enums.js";
+import { type ValueOf } from "~/libs/types/types.js";
+
+const ROUTES_WITHOUT_AUTH_WRAPPER = [AppRoute.QUIZ] as ValueOf<
+	typeof AppRoute
+>[];
+
+export { ROUTES_WITHOUT_AUTH_WRAPPER };
