+++ apps/frontend/src/libs/components/protected-route/libs/helpers/has-auth-wrapper.helper.ts
@@ -1,10 +1,10 @@
-import { type AppRoute } from "~/libs/enums/app-route.enum.js";
+import { type AppRoute } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
-import { routesWithoutAuthWrapper } from "../constants/constants.js";
+import { ROUTES_WITHOUT_AUTH_WRAPPER } from "../constants/constants.js";
 
 function hasAuthWrapper(pathname: ValueOf<typeof AppRoute>): boolean {
-	return !routesWithoutAuthWrapper.includes(pathname);
+	return !ROUTES_WITHOUT_AUTH_WRAPPER.includes(pathname);
 }
 
 export { hasAuthWrapper };
