+++ apps/frontend/src/libs/components/protected-route/libs/helpers/helpers.ts
@@ -1,0 +1,1 @@
+export { hasAuthWrapper } from "./has-auth-wrapper.helper.js";
